Warn when `hashable` is used on an encrypted image
An image's hash is calculated from the following two inputs:
1. Image header.
2. Unencrypted image body.

The `image hashable` command extracts the header and body from a
complete image.  If the image body is encrypted, then the result of this
command will not generate the correct hash.

This commit adds a warning in case `image hashable` is used on an
encrypyted image.